+++ smsapi/smsapi.go
@@ -15,7 +15,7 @@ import (
 )
 
 const (
-	Version        = "1.0.0"
+	Version        = "1.1.0"
 	Name           = "smsapi-go"
 	BaseUrlPl      = "https://api.smsapi.pl/"
 	BaseUrlCom     = "https://api.smsapi.com/"