+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changes
 
+## 1.1.0
+- Page iteration
+- Add page iteration support for blacklist and contacts collection
+- Blacklist fix data type for ExpireAt field
+
 ## 1.0.1
 - Support blacklist API